Remove unused files and update Searchbar component
