+++ app/Admin/Config/page.tsx
@@ -1,122 +1,103 @@
 "use client";
-import ConfigBoard from "@/app/components/ConfigBoard/ConfigBoard";
+import ConfigBoard from "@/app/components/AddConfigPlate/AddConfigPlate";
 import ConfigItem from "@/app/components/ConfigItem/ConfigItem";
 import Pagination from "@/app/components/Pagination/Pagination";
 import Searchbar from "@/app/components/Searchbar/Searchbar";
 import SpecialTxt from "@/app/components/SpecialTxt";
 import { useConfig } from "@/app/util/functions";
-import { AddContestAction, AddContestState, Clan } from "@/app/util/types";
+
 import { Box, Center, Divider, Heading, useToast } from "@chakra-ui/react";
 import React, { useMemo, useReducer, useState } from "react";
 import styles from "./Config.module.css";
 
 // for the button:
-function reduceAddContest(
-  state: AddContestState,
-  action: AddContestAction
-): AddContestState {
-  switch (action.type) {
-    case "UPDATE":
-      return { ...state, [action.field]: action.value };
-    case "RESET":
-      return { Team1: "BW", Team2: "RG", ContestCode: "" };
-    default:
-      return state;
-  }
-}
 
 export default function Config() {
-  const toast = useToast();
+	const toast = useToast();
 
-  const tableCols = useMemo(
-    () => ["Team1", "Team2", "Contest Code", "Date Added", "Live", "Remove"],
-    []
-  );
+	const tableCols = useMemo(
+		() => [
+			"Team1",
+			"Team2",
+			"Contest Code",
+			"Date Added",
+			"Live",
+			"Remove",
+		],
+		[]
+	);
 
-  const defaultContest: AddContestState = {
-    Team1: "BW",
-    Team2: "PP",
-    ContestCode: "",
-  };
+	const [isAddLoading, setIsAddLoading] = useState<boolean>(false);
 
-  const [isAddLoading, setIsAddLoading] = useState<boolean>(false);
+	const maxResults = 5;
+	const [page, setPage] = useState(1);
+	const [searchQuery, setSearchQuery] = useState("");
 
-  const [newContest, dispatchContest] = useReducer(
-    reduceAddContest,
-    defaultContest
-  );
+	const { contests, isLoading, isError, mutate } = useConfig(
+		page,
+		searchQuery,
+		maxResults
+	);
+	let contestNodes;
 
-  const maxResults = 5;
-  const [page, setPage] = useState(1);
-  const [searchQuery, setSearchQuery] = useState("");
+	if (isLoading) contestNodes = <Center>Loading...</Center>;
+	else if (isError) contestNodes = <Center>Error...</Center>;
+	else if (contests) {
+		contestNodes = contests.map((contest) => (
+			<ConfigItem
+				key={String(contest._id!)}
+				mutate={mutate}
+				itemData={contest}
+			/>
+		));
 
-  const { contests, isLoading, isError, mutate } = useConfig(
-    page,
-    searchQuery,
-    maxResults
-  );
-  let contestNodes;
+		if (contestNodes.length === 0) {
+			contestNodes = <Center>No contests found</Center>;
+		}
+	}
 
-  if (isLoading) contestNodes = <Center>Loading...</Center>;
-  else if (isError) contestNodes = <Center>Error...</Center>;
-  else if (contests) {
-    contestNodes = contests.map((contest) => (
-      <ConfigItem
-        key={String(contest._id!)}
-        mutate={mutate}
-        itemData={contest}
-      />
-    ));
+	return (
+		<main className={styles.config}>
+			<Heading fontSize="32px" marginBlock={"32px"}>
+				Active Contests
+			</Heading>
 
-    if (contestNodes.length === 0) {
-      contestNodes = <Center>No contests found</Center>;
-    }
-  }
+			{/* Searchbar here */}
+			<div className={styles.search}>
+				<Searchbar
+					searchQuery={searchQuery}
+					setSearchQuery={setSearchQuery}
+					setPage={setPage}
+				/>
+			</div>
 
-  return (
-    <main className={styles.config}>
-      <Heading fontSize="32px" marginBlock={"32px"}>
-        Active Contests
-      </Heading>
+			{/* form for making the item */}
+			<div className={styles.configBoard}>
+				<div className={styles.header}>
+					{tableCols.map((col) => (
+						<SpecialTxt key={col}>{col}</SpecialTxt>
+					))}
+				</div>
 
-      {/* Searchbar here */}
-      <div className={styles.search}>
-        <Searchbar
-          searchQuery={searchQuery}
-          setSearchQuery={setSearchQuery}
-          setPage={setPage}
-        />
-      </div>
+				<Divider variant="default" />
 
-      {/* form for making the item */}
-      <div className={styles.configBoard}>
-        <div className={styles.header}>
-          {tableCols.map((col) => (
-            <SpecialTxt key={col}>{col}</SpecialTxt>
-          ))}
-        </div>
+				<ConfigBoard
+					toast={toast}
+					isLoading={isAddLoading}
+					setIsLoading={setIsAddLoading}
+					mutate={mutate}
+					setPage={setPage}
+				/>
 
-        <Divider variant="default" />
+				{contestNodes}
+			</div>
 
-        <ConfigBoard
-          toast={toast}
-          isLoading={isAddLoading}
-          setIsLoading={setIsAddLoading}
-          newContest={newContest}
-          dispatchContest={dispatchContest}
-          mutate={mutate}
-          setPage={setPage}
-        />
-
-        {contestNodes}
-      </div>
-
-      <Pagination
-        page={page}
-        setPage={setPage}
-        items={contests}
-        maxResults={maxResults}
-      />
-    </main>
-  );
+			<Pagination
+				page={page}
+				setPage={setPage}
+				items={contests}
+				maxResults={maxResults}
+			/>
+		</main>
+	);
 }
