+++ app/components/AddConfigPlate/index.tsx
@@ -1,0 +1,1 @@
+export { default as default } from "./AddConfigPlate";
