+++ app/components/CustomSelect/index.tsx
@@ -1,0 +1,1 @@
+export { default as default } from "./CustomSelect";
