+++ app/components/Searchbar/Searchbar.tsx
@@ -3,24 +3,24 @@ import React from "react";
 import styles from "./Searchbar.module.css";
 
 type searchbarProps = {
-  searchQuery: string;
-  setSearchQuery: React.Dispatch<React.SetStateAction<string>>;
-  setPage: React.Dispatch<React.SetStateAction<number>>;
+	searchQuery: string;
+	setSearchQuery: React.Dispatch<React.SetStateAction<string>>;
+	setPage: React.Dispatch<React.SetStateAction<number>>;
 };
 
 function Searchbar({ searchQuery, setSearchQuery, setPage }: searchbarProps) {
-  return (
-    <div className={styles.main}>
-      <Input
-        variant={"default"}
-        placeholder="Search"
-        onChange={(e) => {
-          setSearchQuery(e.target.value);
-          setPage(1);
-        }}
-      />
-    </div>
-  );
+	return (
+		<div className={styles.main}>
+			<Input
+				variant={"default"}
+				placeholder="Search"
+				onChange={(e) => {
+					setSearchQuery(e.target.value);
+					setPage(1);
+				}}
+			/>
+		</div>
+	);
 }
 
 export default Searchbar;
